requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==40.0.2
cssselect2==0.7.0
Django==4.2
django-crispy-forms==2.0
fonttools==4.39.3
fpdf==1.7.2
html5lib==1.1
idna==3.4
lxml==4.9.2
oscrypto==1.3.0
Pillow==9.5.0
pycparser==2.21
pydyf==0.6.0
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
pypdf==3.7.1
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.12
requests==2.28.2
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
typing-extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uritools==4.0.1
urllib3==1.26.15
webencodings==0.5.1
xhtml2pdf==0.2.9
zopfli==0.2.2
